Refactor phase4 pipeline into modules
